satinel-processor/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
jinja2
numpy
pillow
rasterio
geopandas
shapely
sentinelhub
segment-geospatial
torch
scipy
opencv-python-headless
# optional later: ultralytics
usgs-api
